Add minimum viable prototybe of the app
Add minimum viable prototybe of the Telephant app.

It is based on:
* FastAPI + Uvicorn for API and web interface server
* Alpine.js + Bootstrap 5.3 for web interface
* SQLLite3 + SQLModel + SQLAlchemy for metadata storage
* plain text / YAML files for data storage

The app supports Google login and own self-managed user (e-mail) +
password auth but no registration. This is meant for internal
self-hosted instances that can be isolated and independent of Google.
Otherwise Google auth is preferred.

Each user has to generate one or more API keys so they can use Telephant
to upload reports.

All reports are public.